# docs/contracts/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "Agencia API"
  "description": "API de la agencia. JSON en snake_case. Los errores usan el esquema Error. Regenerada desde FastAPI."
  "version": "0.1.0"
"paths":
...
  "/api/public/diagnostico":
    "post":
      "tags":
        - "public"
      "summary": "Diagnostico"
      "operationId": "diagnostico_api_public_diagnostico_post"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/DiagnosticoIn"
        "required": true
      "responses":
        "202":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Accepted"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/contacto":
    "post":
      "tags":
        - "public"
      "summary": "Contacto"
      "operationId": "contacto_api_public_contacto_post"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/ContactoIn"
        "required": true
      "responses":
        "202":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Accepted"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/checkout":
    "post":
      "tags":
        - "public"
      "summary": "Checkout Route"
      "operationId": "checkout_route_api_public_checkout_post"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/CheckoutIn"
        "required": true
      "responses":
        "201":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CheckoutOut"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/baja":
    "get":
      "tags":
        - "public"
      "summary": "Baja Get"
      "operationId": "baja_get_api_public_baja_get"
      "parameters":
        -
          "name": "token"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema": {}
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
    "post":
      "tags":
        - "public"
      "summary": "Baja Post"
      "operationId": "baja_post_api_public_baja_post"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/BajaIn"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties":
                  "type": "boolean"
                "title": "Response Baja Post Api Public Baja Post"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/u/{token}":
    "post":
      "tags":
        - "public"
      "summary": "One Click"
      "operationId": "one_click_u__token__post"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties":
                  "type": "boolean"
                "title": "Response One Click U  Token  Post"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/derechos":
    "post":
      "tags":
        - "public"
      "summary": "Derechos"
      "operationId": "derechos_api_public_derechos_post"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/DerechosIn"
        "required": true
      "responses":
        "202":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Accepted"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/demo/{token}":
    "get":
      "tags":
        - "public"
      "summary": "Demo"
      "operationId": "demo_demo__token__get"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema": {}
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/proyecto/{token}":
    "get":
      "tags":
        - "public"
      "summary": "Proyecto"
      "operationId": "proyecto_api_public_proyecto__token__get"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProjectPublic"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/proyecto/{token}/intake":
    "post":
      "tags":
        - "public"
      "summary": "Intake"
      "operationId": "intake_api_public_proyecto__token__intake_post"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": true
              "title": "Payload"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties":
                  "type": "boolean"
                "title": "Response Intake Api Public Proyecto  Token  Intake Post"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/proyecto/{token}/feedback":
    "post":
      "tags":
        - "public"
      "summary": "Feedback"
      "operationId": "feedback_api_public_proyecto__token__feedback_post"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/FeedbackIn"
      "responses":
        "202":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties":
                  "type": "string"
                "title": "Response Feedback Api Public Proyecto  Token  Feedback Post"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/public/proyecto/{token}/aprobar":
    "post":
      "tags":
        - "public"
      "summary": "Aprobar"
      "operationId": "aprobar_api_public_proyecto__token__aprobar_post"
      "parameters":
        -
          "name": "token"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "title": "Token"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties":
                  "type": "string"
                "title": "Response Aprobar Api Public Proyecto  Token  Aprobar Post"
        "422":
          "description": "Validation Error"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HTTPValidationError"
...
"components":
  "schemas":
    "Accepted":
      "properties":
        "id":
          "type": "string"
          "title": "Id"
        "status":
          "type": "string"
          "title": "Status"
      "type": "object"
      "required":
        - "id"
        - "status"
      "title": "Accepted"
...
    "BajaIn":
      "properties":
        "token":
          "type": "string"
          "title": "Token"
      "type": "object"
      "required":
        - "token"
      "title": "BajaIn"
    "CheckoutIn":
      "properties":
        "package_code":
          "type": "string"
          "title": "Package Code"
        "lead_id":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Lead Id"
      "type": "object"
      "required":
        - "package_code"
      "title": "CheckoutIn"
    "CheckoutOut":
      "properties":
        "order_id":
          "type": "string"
          "title": "Order Id"
        "checkout_url":
          "type": "string"
          "title": "Checkout Url"
      "type": "object"
      "required":
        - "order_id"
        - "checkout_url"
      "title": "CheckoutOut"
    "ContactoIn":
      "properties":
        "name":
          "type": "string"
          "title": "Name"
        "email":
          "type": "string"
          "title": "Email"
        "message":
          "type": "string"
          "title": "Message"
        "consent":
          "type": "boolean"
          "title": "Consent"
        "honeypot":
          "type": "string"
          "title": "Honeypot"
          "default": ""
        "turnstile_token":
          "type": "string"
          "title": "Turnstile Token"
          "default": ""
      "type": "object"
      "required":
        - "name"
        - "email"
        - "message"
        - "consent"
      "title": "ContactoIn"
...
    "DerechosIn":
      "properties":
        "kind":
          "type": "string"
          "title": "Kind"
        "email":
          "type": "string"
          "title": "Email"
        "details":
          "type": "string"
          "title": "Details"
        "honeypot":
          "type": "string"
          "title": "Honeypot"
          "default": ""
        "turnstile_token":
          "type": "string"
          "title": "Turnstile Token"
          "default": ""
      "type": "object"
      "required":
        - "kind"
        - "email"
        - "details"
      "title": "DerechosIn"
    "DiagnosticoIn":
      "properties":
        "business":
          "type": "string"
          "title": "Business"
        "email":
          "type": "string"
          "title": "Email"
        "commune":
          "type": "string"
          "title": "Commune"
        "category":
          "type": "string"
          "title": "Category"
        "consent":
          "type": "boolean"
          "title": "Consent"
        "consent_text":
          "type": "string"
          "title": "Consent Text"
        "website_url":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Website Url"
        "honeypot":
          "type": "string"
          "title": "Honeypot"
          "default": ""
        "turnstile_token":
          "type": "string"
          "title": "Turnstile Token"
          "default": ""
      "type": "object"
      "required":
        - "business"
        - "email"
        - "commune"
        - "category"
        - "consent"
        - "consent_text"
      "title": "DiagnosticoIn"
...
    "FeedbackIn":
      "properties":
        "text":
          "type": "string"
          "title": "Text"
      "type": "object"
      "required":
        - "text"
      "title": "FeedbackIn"
    "HTTPValidationError":
      "properties":
        "detail":
          "items":
            "$ref": "#/components/schemas/ValidationError"
          "type": "array"
          "title": "Detail"
      "type": "object"
      "title": "HTTPValidationError"
...
    "ProjectPublic":
      "properties":
        "status":
          "type": "string"
          "title": "Status"
        "business":
          "type": "string"
          "title": "Business"
        "revisions_used":
          "type": "integer"
          "title": "Revisions Used"
        "max_revisions":
          "type": "integer"
          "title": "Max Revisions"
        "preview_url":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Preview Url"
      "type": "object"
      "required":
        - "status"
        - "business"
        - "revisions_used"
        - "max_revisions"
      "title": "ProjectPublic"
...
    "ValidationError":
      "properties":
        "loc":
          "items":
            "anyOf":
              -
                "type": "string"
              -
                "type": "integer"
          "type": "array"
          "title": "Location"
        "msg":
          "type": "string"
          "title": "Message"
        "type":
          "type": "string"
          "title": "Error Type"
        "input":
          "title": "Input"
        "ctx":
          "type": "object"
          "title": "Context"
      "type": "object"
      "required":
        - "loc"
        - "msg"
        - "type"
      "title": "ValidationError"
"servers":
  -
    "url": "http://localhost"